requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
boto3==1.38.13
websockets==15.0.1
googlesearch-python==1.3.0
beautifulsoup4==4.13.4
requests==2.32.3